feat: enhance user registration and profile management
- Added email existence check during user registration to prevent duplicates.
- Updated user profile update logic to check for existing usernames and emails.
- Changed user ID type from uint to int64 in various handlers and models for consistency.
- Modified quote contributor ID type from uint to int64 to accommodate anonymous submissions.
- Adjusted leaderboard logic to handle anonymous users and their quote counts.
- Updated database import logic to reflect changes in contributor ID type.

+++ backend/setup/importer.go
@@ -165,7 +165,7 @@ func importFile(filename string) (imported, skipped int, err error) {
 			quote.UUID = entry.UUID
 		}
 		if entry.CreatorID > 0 {
-			quote.ContributorID = uint(entry.CreatorID)
+			quote.ContributorID = int64(entry.CreatorID)
 		}
 
 		if err := database.DB.Create(&quote).Error; err != nil {
